[64-bit ints] Change `CScriptNum::getint` to return `int64_t`
Summary:
As part of the effort to upgrade Script integers to 64-bits, we upgrade the code to handle `CScriptNum::getint` to return `int64_t`.

There's only a few usages of that function anyway:
- In core_write.cpp, to write script assembly, and tinyformat can handle int64_t for %d
- In interpreter.cpp, to convert to integers (e.g. for OP_ROLL). Here we're just careful to avoid overflows.
- In scriptnum_tests.cpp, to compare to ScriptNum10. This test is now remarkably outdated (almost 9 years, looking at the copyright?), and doesn't really test something we worry about anymore. Instead, we remove CScriptNum10 and change the tests to actually test the behavior of CScriptNum.

(NB: BCH added getint32 and getint64 for the same change, but this is unnecessary and it's cleaner to just stick to one function)

Depends on D17290.

Test Plan: `ninja check`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D16996

+++ src/script/interpreter.cpp
@@ -583,9 +583,9 @@ bool ScriptInterpreter::RunNextOp() {
                     return set_error(serror,
                                      ScriptError::INVALID_STACK_OPERATION);
                 }
-                int n = CScriptNum(stacktop(-1), fRequireMinimal).getint();
+                int64_t n = CScriptNum(stacktop(-1), fRequireMinimal).getint();
                 popstack(stack);
-                if (n < 0 || n >= (int)stack.size()) {
+                if (n < 0 || uint64_t(n) >= uint64_t(stack.size())) {
                     return set_error(serror,
                                      ScriptError::INVALID_STACK_OPERATION);
                 }
@@ -998,7 +998,7 @@ bool ScriptInterpreter::RunNextOp() {
                     return set_error(serror,
                                      ScriptError::INVALID_STACK_OPERATION);
                 }
-                const int nKeysCount =
+                const int64_t nKeysCount =
                     CScriptNum(stacktop(-idxKeyCount), fRequireMinimal)
                         .getint();
                 if (nKeysCount < 0 || nKeysCount > MAX_PUBKEYS_PER_MULTISIG) {
@@ -1018,7 +1018,7 @@ bool ScriptInterpreter::RunNextOp() {
                     return set_error(serror,
                                      ScriptError::INVALID_STACK_OPERATION);
                 }
-                const int nSigsCount =
+                const int64_t nSigsCount =
                     CScriptNum(stacktop(-idxSigCount), fRequireMinimal)
                         .getint();
                 if (nSigsCount < 0 || nSigsCount > nKeysCount) {
